refactor: Fireball Fly: Removed Backwards Option from Custom Technique
