+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/movement/fly/modes/fireball/technieques/FlyFireballLegitTechnique.kt
@@ -25,12 +25,17 @@ import net.ccbluex.liquidbounce.config.ChoiceConfigurable
 import net.ccbluex.liquidbounce.config.ToggleableConfigurable
 import net.ccbluex.liquidbounce.event.*
 import net.ccbluex.liquidbounce.event.events.MovementInputEvent
+import net.ccbluex.liquidbounce.event.events.SimulatedTickEvent
 import net.ccbluex.liquidbounce.features.module.modules.movement.fly.ModuleFly
 import net.ccbluex.liquidbounce.features.module.modules.movement.fly.modes.fireball.FlyFireball
 import net.ccbluex.liquidbounce.features.module.modules.movement.fly.modes.fireball.trigger.FlyFireballInstantTrigger
 import net.ccbluex.liquidbounce.features.module.modules.movement.fly.modes.fireball.trigger.FlyFireballOnEdgeTrigger
+import net.ccbluex.liquidbounce.utils.aiming.Rotation
+import net.ccbluex.liquidbounce.utils.aiming.RotationManager
+import net.ccbluex.liquidbounce.utils.aiming.RotationsConfigurable
 import net.ccbluex.liquidbounce.utils.entity.isCloseToEdge
 import net.ccbluex.liquidbounce.utils.kotlin.EventPriorityConvention
+import net.ccbluex.liquidbounce.utils.kotlin.Priority
 import net.ccbluex.liquidbounce.utils.movement.DirectionalInput
 import kotlin.coroutines.suspendCoroutine
 
@@ -48,8 +53,23 @@ object FlyFireballLegitTechnique : Choice("Legit") {
 
     var canMove = true
 
+    object Rotations : RotationsConfigurable(80f..120f) {
+        val pitch by float("Pitch", 90f, 0f..90f)
+        val backwards by boolean("Backwards", true)
+    }
+
     init {
         tree(Jump)
+        tree(Rotations)
+    }
+
+    private val rotationUpdateHandler = handler<SimulatedTickEvent> {
+        RotationManager.aimAt(
+            Rotation(if (Rotations.backwards) RotationManager.invertYaw(player.yaw) else player.yaw, Rotations.pitch),
+            configurable = Rotations,
+            priority = Priority.IMPORTANT_FOR_PLAYER_LIFE,
+            provider = ModuleFly
+        )
     }
 
     private val movementInputHandler = sequenceHandler<MovementInputEvent> { event ->
