+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/world/scaffold/ModuleScaffold.kt
@@ -99,7 +99,7 @@ object ModuleScaffold : Module("Scaffold", Category.WORLD) {
         val slotResetDelay by int("SlotResetDelay", 5, 0..40, "ticks")
     }
 
-    init {ModuleScaffold
+    init {
         tree(AutoBlock)
     }
